pyproject.toml
[tool.poetry]
name = "newsbot"
version = "0.1.0"
description = "Telegram bot aggregator for financial news"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.1"
openai = "^1.23.0"
google-generativeai = "^0.5.1"
langchain = "^0.1.17"
guardrails-ai = "^0.3.3"
tinkoff-investments = "^2.0.1"
feedparser = "^6.0.11"
aiohttp = "^3.9.5"
SQLAlchemy = {version="^2.0.30", extras=["asyncio"]}
asyncpg = "^0.29.0"
apscheduler = "^3.10.4"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.5"
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.ruff]
line-length = 88

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
